feat(mobile): update asset detail header typography to md variant
- Replace custom Trend component with TrendSection wrapper using Lumen Trend (size="md")
- Update "Market price" subtitle from body3 to body2
- Align formattedChange and timeLabel to same muted color
- Add NaN/0 guard in TrendSection

LIVE-30941
