+++ .changeset/live-30941-header-typography-adjustment.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Update asset detail header typography: "Market price" subtitle and trend to MD variant; replace custom Trend with Lumen Trend component via TrendSection wrapper
